ml/pytorch/requirements.txt
python-dotenv
torch
torchvision
